Fixed column which mixed strings with integers
Some of the values over 1000 in the votes column were being listed inside double quotes with a thousands comma separator. This meant that the overall column is treated by some tools as a string, not an integer. Cleaned up that column so that tools would detect the column as being of integer type.